+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file. The format 
 
 ### Added
 
+- Add a `random` utility function ([#662](https://github.com/studiometa/js-toolkit/pull/662), [a25247a5](https://github.com/studiometa/js-toolkit/commit/a25247a5))
 - Add a `smoothTo` utility function ([#660](https://github.com/studiometa/js-toolkit/pull/660), [f4d7b349](https://github.com/studiometa/js-toolkit/commit/f4d7b349))
 - Add a `wrap` utility function ([#660](https://github.com/studiometa/js-toolkit/pull/660), [866ed3a4](https://github.com/studiometa/js-toolkit/commit/866ed3a4))
 - Add a `spring` function ([#660](https://github.com/studiometa/js-toolkit/pull/660), [1e06f792](https://github.com/studiometa/js-toolkit/commit/1e06f792))